+++ app/client/components/helmet/types.js
@@ -1,0 +1,2 @@
+export type UpdateTitleType = (title: string) => { payload: string };
+export type UpdateLinkType = (link: Object[]) => { payload: Object[] };