+++ app/client/components/todos/index.jsx
@@ -3,13 +3,14 @@ import React from 'react';
 import { compose, onlyUpdateForKeys } from 'recompose';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import fetchDataEnhancer from './../../helpers/fetch-data-enhancer';
 import TodosHeader from './todos-header';
 import TodosAdd from './todos-add';
 import TodosBody from './todos-body';
 import TodosFooter from './todos-footer';
-import { selectors, addTodo, removeTodo, completeTodo, fetchTodos } from './logic-bundle';
 import type { TodoType } from './types';
+import { selectors, addTodo, removeTodo, completeTodo, fetchTodos } from './logic-bundle';
+import { updateLink } from './../helmet/logic-bundle';
+import redialEnhancer, { FETCH_DATA_HOOK, INJECT_PRELOAD_LINK_HOOK } from './../../helpers/redial-enhancer';
 
 export const Todos = ({ todos, actions }: { todos: TodoType[], actions: Object }) => (
   <div className="container">
@@ -27,9 +28,13 @@ export const Todos = ({ todos, actions }: { todos: TodoType[], actions: Object }
 );
 
 export const enhance = compose(
-  fetchDataEnhancer(
-    ({ store }) => store.dispatch(fetchTodos())
-  ),
+  redialEnhancer({
+    [FETCH_DATA_HOOK]: ({ store }) => store.dispatch(fetchTodos()),
+    [INJECT_PRELOAD_LINK_HOOK]: ({ store }) => store.dispatch(updateLink([
+      // window.javascriptAssets will be injected to do preload link for optimizing route
+      { rel: 'preload', href: window.javascriptAssets['static-page'], as: 'script' },
+    ])),
+  }),
   connect(
     state => ({
       todos: selectors.getTodos(state),